Revert "feat: support streaming with knowledge base (PL-987) (#814)"
