--- lib/services/aiSynthesis/buffer-reduce.exception.ts
@@ -1,1 +1,0 @@
-export class BufferedReducerStopException extends Error {}
