--- lib/services/aiSynthesis/types.ts
@@ -1,4 +1,0 @@
-import type { AIResponse } from '@/lib/services/runtime/handlers/utils/ai';
-import type { KnowledgeBaseFaqSet, KnowledgeBaseResponse } from '@/lib/services/runtime/handlers/utils/knowledgeBase';
-
-export type KBResponse = AIResponse & Partial<KnowledgeBaseResponse> & { faqSet?: KnowledgeBaseFaqSet };
