+++ tests/lib/services/aiSynthesis.unit.ts
